requirements.txt
acres==0.3.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.17
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
arxiv==2.2.0
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
comm==0.2.2
configobj==5.0.9
configparser==7.2.0
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
dnspython==2.7.0
email_validator==2.2.0
etelemetry==0.3.1
executing==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
feedparser==6.0.11
filelock==3.18.0
frozenlist==1.6.0
greenlet==3.2.0
h11==0.14.0
httpcore==1.0.8
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
isodate==0.6.1
jedi==0.19.2
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.54
langchain-text-splitters==0.3.8
langsmith==0.3.32
looseversion==1.3.0
lxml==5.3.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.4.3
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
numpy==2.2.5
orjson==3.10.16
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathlib==1.0.1
pexpect==4.9.0
platformdirs==4.3.7
prompt_toolkit==3.0.51
propcache==0.3.1
prov==2.0.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.28
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
pydot==3.0.4
Pygments==2.19.1
PyMuPDF==1.25.5
pyparsing==3.2.3
pypdf==5.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
pyzmq==26.4.0
rdflib==6.3.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.1
scipy==1.15.2
sgmllib3k==1.0.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
stack-data==0.6.3
starlette==0.46.2
tenacity==9.1.2
tornado==6.4.2
traitlets==5.14.3
traits==7.0.2
typer==0.15.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
yarl==1.20.0
zstandard==0.23.0

langchain-groq>=0.0.7
pymongo>=4.6.0
chromadb>=0.4.22
sentence-transformers>=2.5.0
langchain-huggingface>=0.0.1
langchain-chroma>=0.0.1 
